# src/main/resources/com/sonicle/webtop/calendar/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Calendar",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Calendar service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    },
    {
      "auth-apikey-bearer": [],
      "auth-apikey-username": []
    }
  ],
  "paths": {
...
    "/me/calendars": {
      "get": {
        "summary": "List Calendars",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarsResult"
                }
              }
            }
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "operationId": "listCalendars",
        "description": "Returns a list of Calendars readable by the current user: this includes both personal and incoming shared categories.\n\nParam *_filter* supports the following fields in RSQL condition:\nid, createdAt, updatedAt, userId, builtIn, provider, name, description, color, easSync\n\nParam *_select* supports the following fields:\nprovider, builtIn, name, description, color, easSync, defSensitivity, defShowAs, defReminder, remoteSyncFrequency, remoteSyncTimestamp, remoteSyncToken",
        "parameters": [
          {
            "$ref": "#/components/parameters/_filter"
          },
          {
            "$ref": "#/components/parameters/_select"
          },
          {
            "$ref": "#/components/parameters/_orderBy"
          },
          {
            "$ref": "#/components/parameters/_pageNo"
          },
          {
            "$ref": "#/components/parameters/_pageSize"
          },
          {
            "$ref": "#/components/parameters/_returnCount"
          }
        ]
      },
      "post": {
        "summary": "Add a Calendar",
        "operationId": "addCalendar",
        "responses": {
          "201": {
            "description": "Calendar created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "tags": [
          "me"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "user_id",
            "description": "The ID of a user"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarBase"
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Adds new category specifying the owning user ID. If no user ID is provided, the owner will be the current user."
      }
    },
    "/me/calendars/{calendar_id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "calendar_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get a Calendar",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Calendar not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "operationId": "getCalendar",
        "description": "Gets the specified Calendar given its ID."
      },
      "put": {
        "summary": "Update a Calendar",
        "operationId": "updateCalendar",
        "responses": {
          "200": {
            "description": "Calendar updated"
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Calendar not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "tags": [
          "me"
        ],
        "x-codegen-request-body-name": "body",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarBase"
              }
            }
          }
        },
        "description": "Update the specified Calendar given its ID. You can choose to update the entire object or only a subset of data."
      },
      "delete": {
        "summary": "Delete a Calendar",
        "operationId": "deleteCalendar",
        "responses": {
          "204": {
            "description": "Calendar deleted"
          },
          "400": {
            "description": "Invalid parameter"
          },
          "404": {
            "description": "Calendar not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "tags": [
          "me"
        ],
        "description": "Delete a Calendar given its ID."
      }
    },
    "/me/calendars/{calendar_id}/events": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "calendar_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "List Events of a Calendar",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResult"
                }
              }
            }
          },
          "404": {
            "description": "Calendar not found"
          },
          "405": {
            "description": "Not allowed"
          }
        },
        "operationId": "listCalendarEvents",
        "parameters": [
          {
            "$ref": "#/components/parameters/_filter"
          },
          {
            "$ref": "#/components/parameters/_select"
          },
          {
            "$ref": "#/components/parameters/_orderBy"
          },
          {
            "$ref": "#/components/parameters/_pageNo"
          },
          {
            "$ref": "#/components/parameters/_pageSize"
          },
          {
            "$ref": "#/components/parameters/_returnCount"
          }
        ],
        "description": "Returns a list of Events from specified Calendar.\n\nParam *_filter* supports the following fields in RSQL query conditions:\nid, createdAt, updatedAt, rowStatus, status, organizer, organizerId, title, location, description, timezone, allDay, start, end, visibility, transparency, tagId\n\nParam *_select* supports the following fields:\npublicUid, location, description, visibility, transparency, href, reminder"
      },
      "post": {
        "summary": "Add an Event into Calendar",
        "tags": [
          "me"
        ],
        "operationId": "addCalendarEvent",
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Calendar not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventEx"
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Adds new Event into given Calendar."
      }
    },
    "/me/calendars/{calendar_id}/events/delta": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "calendar_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "List changes on Events collection",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResultDelta"
                }
              }
            }
          }
        },
        "operationId": "listCalendarEventsDelta",
        "parameters": [
          {
            "$ref": "#/components/parameters/_syncToken"
          },
          {
            "$ref": "#/components/parameters/_select"
          }
        ],
        "description": "Get a set of Events that have been added, deleted, or updated in a specified Calendar, starting from a precise instant identified by a given syncToken.\n\nParam *_select* supports the following fields:\nlocation, description, visibility, transparency, href, reminder"
      }
    },
...
    "/me/events/instances/{event_instance_id}/response": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "event_instance_id",
          "in": "path",
          "required": true
        }
      ],
      "put": {
        "summary": "Update the response to an Event invitation",
        "operationId": "updateEventInstanceResponse",
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Event or invitation not found"
          }
        },
        "tags": [
          "me"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventResponse"
              }
            }
          }
        },
        "description": "Update the invitation response for the specified Event instance."
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "auth-apikey-bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "APIKey",
        "description": "API-Key Authentication: specifies the secret token (bearer and username are both required)"
      },
      "auth-apikey-username": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Username",
        "description": "API-Key Authentication: specifies target username (bearer and username are both required)"
      }
    },
    "parameters": {
...
      "_select": {
        "name": "_select",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "example": "name, description"
        },
        "description": "List (comma-separated) of field names to include in resulting items. Optional, if omitted all available field will be taken into account."
      },
...
      "_filter": {
        "name": "_filter",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A RSQL filter query to filter out resulting items. Optional."
      },
...
    }
  },
...
}